+++ CHANGELOG.md
@@ -8,6 +8,10 @@ ml).
 ## master
 
 ## testing
+- [added] #219 Add functionality to searchbar (Lots and devices).
+- [changed] #211 Print DHID-QR label for selected devices.
+- [changed] #218 Add reactivity to device lots.
+- [fixed] #214 Login workflow
 
 ## [2.0.0] - 2022-03-15
 First server render HTML version. Completely rewrites views of angular JS client on flask.